Further decompose isSupported
